+++ panel/db.go
@@ -89,6 +89,9 @@ func migratePanelDB() error {
 		if err := migratePanelDBV11(); err != nil {
 			return err
 		}
+		if err := migratePanelDBV12(); err != nil {
+			return err
+		}
 		if _, err = panelDB.Exec(`INSERT INTO schema_version (version) VALUES (?)`, dbSchemaVersion); err != nil {
 			return err
 		}
@@ -145,6 +148,11 @@ func migratePanelDB() error {
 	}
 	if ver < 11 {
 		if err := migratePanelDBV11(); err != nil {
+			return err
+		}
+	}
+	if ver < 12 {
+		if err := migratePanelDBV12(); err != nil {
 			return err
 		}
 	}